requirements.txt
Django==4.2.6
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.0
djangorestframework==3.14.0
idna==3.4
pip==22.3.1
pytz==2023.3.post1
requests==2.31.0
setuptools==65.5.1
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==2.0.6
wheel==0.38.4